+++ lib/excel_plus.dart
@@ -8,7 +8,7 @@ import 'package:archive/archive.dart';
 import 'package:xml/xml.dart';
 import 'package:xml/xml_events.dart';
 import 'src/platform/save_stub.dart'
-    if (dart.library.html) 'src/platform/save_web.dart'
+    if (dart.library.js_interop) 'src/platform/save_web.dart'
     as helper;
 
 /// Core
@@ -52,6 +52,7 @@ part 'src/utils/archive_utils.dart';
 part 'src/utils/cell_utils.dart';
 part 'src/utils/color_utils.dart';
 part 'src/utils/fast_list.dart';
+part 'src/utils/worksheet_order.dart';
 
 XmlName _xmlName(String local, [String? prefix]) =>
     XmlName.parts(local, prefix: prefix);